NATIVE-632: Rename to PushNotificationsManager.js
