+++ src/modules/i18n/watchClearCity.js
@@ -3,7 +3,7 @@
 import type { Saga } from 'redux-saga'
 import { takeLatest, call } from 'redux-saga/effects'
 import AppSettings from '../settings/AppSettings'
-import * as NotificationsManager from '../../modules/notifications/NotificationsManager'
+import * as NotificationsManager from '../push-notifications/PushNotificationsManager'
 import buildConfig from '../app/constants/buildConfig'
 
 export function * clearCity (): Saga<void> {
